Don't re-download the library files if already existing

+++ setup.py
@@ -1,60 +1,47 @@
 #!/usr/bin/env python3
 
 import sys
-from setuptools import setup, find_packages, Command
+from setuptools import setup, find_packages
 import urllib.request
 import shutil
+import os.path
 
 # Package meta-data
 VERSION = '0.22.0'
 CORE_VERSION = '0.22.0'  # graphANNIS core library version
 
 
-class CoreLibraryCommand(Command):
-    """Download released graphANNIS core library binaries"""
-
-    description = 'Download released graphANNIS core library binaries'
-    user_options = []
-
-    @staticmethod
-    def status(s):
-        """Prints things in bold."""
-        print('\033[1m{0}\033[0m'.format(s))
-
-    def initialize_options(self):
-        pass
-
-    def finalize_options(self):
-        pass
-
-    def run(self):
-        self.status('Downloading released graphANNIS ' +
-                    VERSION + ' core library binaries')
-
-        self.status('libgraphannis.so (Linux)')
-        url = 'https://github.com/korpling/graphANNIS/releases/download/v' + CORE_VERSION + '/libgraphannis.so'
-        file = 'graphannis/linux-x86-64/libgraphannis.so'
-        with urllib.request.urlopen(url) as response, open(file, 'wb') as out_file:
-            shutil.copyfileobj(response, out_file)
-
-        self.status('graphannis.dll (Windows)')
-        url = 'https://github.com/korpling/graphANNIS/releases/download/v' + CORE_VERSION + '/graphannis.dll'
-        file = 'graphannis/win32-x86-64/raphannis.dll'
-        with urllib.request.urlopen(url) as response, open(file, 'wb') as out_file:
-            shutil.copyfileobj(response, out_file)
-
-
-        self.status('libgraphannis.dylib (OSX)')
-        url = 'https://github.com/korpling/graphANNIS/releases/download/v' + CORE_VERSION + '/libgraphannis.dylib'
-        file = 'graphannis/darwin-x86-64/libgraphannis.dylib'
-        with urllib.request.urlopen(url) as response, open(file, 'wb') as out_file:
-            shutil.copyfileobj(response, out_file)
-
-        sys.exit()
-
-
 with open('README_pypi.md') as f:
     long_description = f.read()
+
+print('Downloading released graphANNIS ' +
+            VERSION + ' core library binaries')
+
+print('libgraphannis.so (Linux)')
+url = 'https://github.com/korpling/graphANNIS/releases/download/v' + \
+    CORE_VERSION + '/libgraphannis.so'
+file = 'graphannis/linux-x86-64/libgraphannis.so'
+if not os.path.isfile(file):
+    with urllib.request.urlopen(url) as response, open(file, 'wb') as out_file:
+        shutil.copyfileobj(response, out_file)
+
+print('graphannis.dll (Windows)')
+url = 'https://github.com/korpling/graphANNIS/releases/download/v' + \
+    CORE_VERSION + '/graphannis.dll'
+file = 'graphannis/win32-x86-64/raphannis.dll'
+if not os.path.isfile(file):
+    with urllib.request.urlopen(url) as response, open(file, 'wb') as out_file:
+        shutil.copyfileobj(response, out_file)
+
+print('libgraphannis.dylib (OSX)')
+url = 'https://github.com/korpling/graphANNIS/releases/download/v' + \
+    CORE_VERSION + '/libgraphannis.dylib'
+file = 'graphannis/darwin-x86-64/libgraphannis.dylib'
+if not os.path.isfile(file):
+    with urllib.request.urlopen(url) as response, open(file, 'wb') as out_file:
+        shutil.copyfileobj(response, out_file)
+print("Finished downloading core library files")
+
 
 setup(name='graphannis',
       version=VERSION,
@@ -75,8 +62,5 @@ setup(name='graphannis',
           "Operating System :: POSIX :: Linux",
           "Operating System :: MacOS :: MacOS X",
           "Operating System :: Microsoft :: Windows"
-      ),
-      cmdclass={
-          'build_clib': CoreLibraryCommand
-      }
+      )
       )
